Moved UI out of APP
- Had to update Tailwind config to encompass UI folder.
- Also added video loading optimization to BackgroundVideo compoennt
