+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import Header from "./UI/Header/Header";
-import "./globals.css";
+import Header from "@/UI/Header/Header";
+import '../globals.css'
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
