+++ tailwind.config.ts
@@ -5,6 +5,7 @@ export default {
     "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/UI/**/*.{js,ts,jsx,tsx,mdx}",
   ],
   theme: {
     extend: {
